DPTS-824 Wip refactoring of int test code
