+++ test/unit/hk/integration/test_symptomgroup.py
@@ -3,7 +3,7 @@ import pytest
 import string
 import psycopg2
 
-from ..model import symptomgroup
+from ..models import symptomgroup
 
 file_path = "application.hk.integration.model.symptomgroup"
